+++ src/components/Geospatial/Legends/LegendPanel/styled.ts
@@ -1,7 +1,7 @@
 import { css } from '@emotion/react'
 
 export const legendPanelContainerStyles = css`
-  width: 290px;
+  width: 100%;
   box-shadow: 0px 2px 4px -2px #0000001a;
   box-shadow: 0px 4px 6px -1px #0000001a;
 `