+++ src/components/Geospatial/Legends/LegendPanel/types.ts
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export type LegendPanelProps = {
-  legendContent: React.ReactElement[]
-  analysisContent: React.ReactNode
-  onTabClick?: (tabValue: string) => void
-}